+++ src/js/nice-select2.js
@@ -66,6 +66,8 @@ class NiceSelect {
   }
 
   create() {
+    this.el.classList.add('hidden-select');
+    
     /* Object.assign(this.el.style, {
       opacity: "0",
       width: "0",
@@ -109,14 +111,14 @@ class NiceSelect {
           text,
           value: item.value,
           extra: item.dataset.extra,
-          selected: item.hasAttribute("selected"),
-          disabled: item.hasAttribute("disabled"),
+          selected: item.selected,
+          disabled: item.disabled,
         };
       }
 
       const attributes = {
-        selected: item.hasAttribute("selected"),
-        disabled: item.hasAttribute("disabled"),
+        selected: item.selected,
+        disabled: item.disabled,
         optgroup: item.tagName === "OPTGROUP",
       };
 
@@ -303,13 +305,7 @@ class NiceSelect {
 
     if (this.dropdown) {
       this.dropdown.remove();
-      this.el.style.display   = "";
-      this.el.style.opacity   = '';
-      this.el.style.width     = '';
-      this.el.style.padding   = '';
-      this.el.style.height    = '';
-      this.el.style.fontSize  = '';
-      this.el.style.minHeight = '';
+      this.el.classList.remove('hidden-select');
     }
   }